Moves websocket proxy from primary cpyress config to the smoketests config
